// frontend/web/src/services/API/proserve-wb-provisioning-api/models/AdditionalConfiguration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { AdditionalConfigurationParameter } from './AdditionalConfigurationParameter';
import {
    AdditionalConfigurationParameterFromJSON,
    AdditionalConfigurationParameterFromJSONTyped,
    AdditionalConfigurationParameterToJSON,
} from './AdditionalConfigurationParameter';

/**
 * 
 * @export
 * @interface AdditionalConfiguration
 */
export interface AdditionalConfiguration {
    /**
     * The configuration type.
     * @type {string}
     * @memberof AdditionalConfiguration
     */
    type: string;
    /**
     * The configuration parameters.
     * @type {Array<AdditionalConfigurationParameter>}
     * @memberof AdditionalConfiguration
     */
    parameters?: Array<AdditionalConfigurationParameter>;
}

/**
 * Check if a given object implements the AdditionalConfiguration interface.
 */
export function instanceOfAdditionalConfiguration(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "type" in value;

    return isInstance;
}

export function AdditionalConfigurationFromJSON(json: any): AdditionalConfiguration {
    return AdditionalConfigurationFromJSONTyped(json, false);
}

export function AdditionalConfigurationFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdditionalConfiguration {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'type': json['type'],
        'parameters': !exists(json, 'parameters') ? undefined : ((json['parameters'] as Array<any>).map(AdditionalConfigurationParameterFromJSON)),
    };
}

export function AdditionalConfigurationToJSON(value?: AdditionalConfiguration | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'type': value.type,
        'parameters': value.parameters === undefined ? undefined : ((value.parameters as Array<any>).map(AdditionalConfigurationParameterToJSON)),
    };
}