// frontend/web/src/services/API/proserve-wb-provisioning-api/models/AdditionalConfigurationParameter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AdditionalConfigurationParameter
 */
export interface AdditionalConfigurationParameter {
    /**
     * The parameter key.
     * @type {string}
     * @memberof AdditionalConfigurationParameter
     */
    key: string;
    /**
     * The parameter value.
     * @type {string}
     * @memberof AdditionalConfigurationParameter
     */
    value?: string;
}

/**
 * Check if a given object implements the AdditionalConfigurationParameter interface.
 */
export function instanceOfAdditionalConfigurationParameter(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "key" in value;

    return isInstance;
}

export function AdditionalConfigurationParameterFromJSON(json: any): AdditionalConfigurationParameter {
    return AdditionalConfigurationParameterFromJSONTyped(json, false);
}

export function AdditionalConfigurationParameterFromJSONTyped(json: any, ignoreDiscriminator: boolean): AdditionalConfigurationParameter {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'key': json['key'],
        'value': !exists(json, 'value') ? undefined : json['value'],
    };
}

export function AdditionalConfigurationParameterToJSON(value?: AdditionalConfigurationParameter | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'key': value.key,
        'value': value.value,
    };
}